♻️ refactor: data fetch 관련 모듈 modules 폴더에 정리
